# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Movies API 3.0.0",
    "description": "Movies Api",
    "contact": {
      "email": "arthagusfi8@gmail.com"
    },
    "version": "1.0.11"
  },
  "servers": [
    {
      "url": "https://localhost:3000"
    }
  ],
  "paths": {
    "/movies/{id}": {
      "get": {
        "summary": "find a movies by id",
        "tags": [
          "movies"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "schema": {
              "type": "integer"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Movies"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "create a movies by id",
        "tags": [
          "movies"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "schema": {
              "type": "integer"
            },
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful create a movies",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Succes"
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "create a movies by id",
        "tags": [
          "movies"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "schema": {
              "type": "integer"
            },
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Edit a movies",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Succes"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "movies"
        ],
        "summary": "Deletes a movies by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Succes delete movies",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SuccesDelete"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Invalid Id Or Id Does Not Exist"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Movies": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "title": {
            "type": "string",
            "example": "Reckless"
          },
          "genres": {
            "type": "string",
            "example": "Drama|Comedy|Romance"
          },
          "year": {
            "type": "string",
            "example": 2001
          }
        }
      },
      "Succes": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Create a movies, Succes"
          }
        }
      },
      "SuccesDelete": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Succesfuly deleted"
          }
        }
      }
    }
  }
}